docs/requirements-docs.txt
tox
Sphinx
doc8
docutils
setuptools
six
sphinx_rtd_theme>=0.4.3
sphinxcontrib-blockdiag
sphinxcontrib-needs>=0.2.3
sphinxcontrib-nwdiag
sphinxcontrib-seqdiag
sphinxcontrib-swaggerdoc
sphinxcontrib-plantuml
sphinx_bootstrap_theme
lfdocs-conf
